Fix lint/tsc issues when `strict` tsc flag is `true` (#487)
This commit aims to fix any issues when the `strict` flag is on without changing
any of the current behavior, or public api at all. The idea is that no consumer
will be adversely affected by this change. It will allow consumers to opt in
to `strict` without any issues from skyux builder.